+++ src/Kunstmaan/UserManagementBundle/Controller/RolesController.php
@@ -10,11 +10,12 @@ use Kunstmaan\AdminBundle\FlashMessages\FlashTypes;
 use Kunstmaan\AdminBundle\Form\RoleType;
 use Kunstmaan\AdminListBundle\AdminList\AdminList;
 use Kunstmaan\UserManagementBundle\AdminList\RoleAdminListConfigurator;
-use Symfony\Component\HttpFoundation\RedirectResponse;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 /**
  * Settings controller handling everything related to creating, editing, deleting and listing roles in an admin list
@@ -32,7 +33,7 @@ class RolesController extends BaseSettingsController
      */
     public function listAction(Request $request)
     {
-        $this->checkPermission();
+        $this->denyAccessUnlessGranted('ROLE_SUPER_ADMIN');
 
         $em        = $this->getDoctrine()->getManager();
         /* @var AdminList $adminlist */
@@ -56,7 +57,7 @@ class RolesController extends BaseSettingsController
      */
     public function addAction(Request $request)
     {
-        $this->checkPermission();
+        $this->denyAccessUnlessGranted('ROLE_SUPER_ADMIN');
 
         /* @var $em EntityManager */
         $em = $this->getDoctrine()->getManager();
@@ -69,9 +70,11 @@ class RolesController extends BaseSettingsController
                 $em->persist($role);
                 $em->flush();
 
-                $this->get('session')->getFlashBag()->add(
+                $this->addFlash(
                     FlashTypes::SUCCESS,
-                    $this->get('translator')->trans('kuma_user.roles.add.flash.success.%role%', ['%role%' => $role->getRole()])
+                    $this->get('translator')->trans('kuma_user.roles.add.flash.success.%role%', [
+                        '%role%' => $role->getRole()
+                    ])
                 );
 
                 return new RedirectResponse($this->generateUrl('KunstmaanUserManagementBundle_settings_roles'));
@@ -97,7 +100,7 @@ class RolesController extends BaseSettingsController
      */
     public function editAction(Request $request, $id)
     {
-        $this->checkPermission();
+        $this->denyAccessUnlessGranted('ROLE_SUPER_ADMIN');
 
         /* @var $em EntityManager */
         $em = $this->getDoctrine()->getManager();
@@ -111,9 +114,11 @@ class RolesController extends BaseSettingsController
                 $em->persist($role);
                 $em->flush();
 
-                $this->get('session')->getFlashBag()->add(
+                $this->addFlash(
                     FlashTypes::SUCCESS,
-                    $this->get('translator')->trans('kuma_user.roles.edit.flash.success.%role%', ['%role%' => $role->getRole()])
+                    $this->get('translator')->trans('kuma_user.roles.edit.flash.success.%role%', [
+                        '%role%' => $role->getRole()
+                    ])
                 );
 
                 return new RedirectResponse($this->generateUrl('KunstmaanUserManagementBundle_settings_roles'));
@@ -139,20 +144,21 @@ class RolesController extends BaseSettingsController
      */
     public function deleteAction($id)
     {
-        $this->checkPermission();
+        $this->denyAccessUnlessGranted('ROLE_SUPER_ADMIN');
 
         /* @var $em EntityManager */
         $em = $this->getDoctrine()->getManager();
         /* @var Role $role */
         $role = $em->getRepository('KunstmaanAdminBundle:Role')->find($id);
         if (!is_null($role)) {
-            $roleName = $role->getRole();
             $em->remove($role);
             $em->flush();
 
-            $this->get('session')->getFlashBag()->add(
+            $this->addFlash(
                 FlashTypes::SUCCESS,
-                $this->get('translator')->trans('kuma_user.roles.delete.flash.success.%role%', ['%role%' => $roleName])
+                $this->get('translator')->trans('kuma_user.roles.delete.flash.success.%role%', [
+                    '%role%' => $role->getRole()
+                ])
             );
         }
